requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
gradio==4.19.2
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24
langchain-chroma==0.0.6
langchain-text-splitters==0.0.1
sentence-transformers==2.5.1
chromadb==0.4.22
pydantic==2.6.1
python-multipart==0.0.9
psutil==5.9.8  # For memory monitoring 